requirements.txt
fastapi
uvicorn
python-multipart
pyjwt
passlib
PyPDF2
python-jose[cryptography]
passlib[bcrypt]
